fix(grid_tile_item): cellCountWidth changed  to int type

+++ lib/views/common/grid_view_widget/grid_tile_item.dart
@@ -1,4 +1,3 @@
-
 import 'package:flutter/material.dart';
 
 const labelTileStyle = TextStyle(
@@ -15,7 +14,7 @@ class GridTileItem extends StatelessWidget {
   final Widget? child;
   final String? hintText;
   final Color? color;
-  final double? cellCountWidth;
+  final int? cellCountWidth;
 
   const GridTileItem({
     super.key,
